backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==4.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# AI and ML Libraries
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langgraph==0.0.26
openai==1.6.1
tiktoken==0.5.2

# Korean Language Processing
sentence-transformers==2.2.2
kiwipiepy==0.17.0

# Vector Database
pymilvus==2.3.4
transformers==4.36.2
torch>=2.2.0
elasticsearch==8.11.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.0

# Data Validation & Serialization
email-validator==2.1.0
python-dateutil==2.8.2

# File Processing
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2
unstructured==0.11.8
markdown==3.5.1
beautifulsoup4==4.12.2

# File Handling and Storage
aiofiles==24.1.0

# Environment & Configuration
python-dotenv==1.0.0
click==8.1.7
pyyaml==6.0.1

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0
structlog==23.2.0

# Background Tasks
celery[redis]==5.3.4
flower==2.0.1
apscheduler==3.10.4

# WebSocket support
websockets==12.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# API Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0